Extract iPod active library hook
