+++ src/apps/ipod/hooks/index.ts
@@ -1,2 +1,3 @@
 // Re-export all iPod hooks
+export { useIpodActiveLibrary } from "./useIpodActiveLibrary";
 export { useLibraryUpdateChecker } from "./useLibraryUpdateChecker";
